miniboot: fix incorrect vlan interface applied for default route
This commit fixes a but where the ip route add default is referencing
the management device name instead of the interface name (containing
vlan tag).

The issue is only seen when the OAM network is on a VLAN and
is a separate network (requires nexthop_gateway setting in
install values).

The fix is to apply the root on the vlan interface, not the top-level
network device interface.

Test Plan
PASS:
- Verify installation on system with OAM network on separate VLAN using
  nexthop_gateway
- Verify installation on system with vlan but no nexthop_gateway

Closes-Bug: 2013372
Change-Id: Ic3febbd0cb77dd21435f23859e6d228e6ab95a8c
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>