Add intel multi-drivers-switch kernel parameter support to kickstarts
This update adds support to the Debian kickstarts to search the
install kernel command line for the multi-drivers-switch= option.
If that option is found, then the full option with the specified
version, ex: multi-drivers-switch=2.54 , will be added to the
disk boot kernel command line options.

Test Plan:

PASS: Verify Build and Install SX system
PASS: Unit test of code block function over an install
PASS: Verify if the multi-drivers-switch parameter exists on the
      the node install command line then the same option is
      propagated to the disk boot command line.
PASS: Verify the opposite of the above is true.

Closes-Bug: 2026893
Change-Id: I648b16dbc5aa2a0a7b8368c1b89a5d46418ab1e5
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>