Fix ip -6 address netmask and workaround for multi-drivers-switch
This commit fixes the missing support for bootstrap_address_prefix
in the miniboot ip -6 address add command. We check for the provided
prefix value parsed from the boot arguments and make sure that it is
applied if present. Note that the bootstrap_address_prefix is a
mandatory install value, so it will be provided. However, we leave the
capability for it to be missing, in order to de-risk this commit.

Additionally, a workaround is included for full support of
multi-drivers-switch given in the boot arguments. When this argument is
given we parse out the kernel module version and use it to replace the
current kernel modules for ice/i40e/iavf with the modules of the given
version.

Test Plan
PASS:
- Replace  miniboot.cfg at /var/miniboot/kickstart-override/miniboot.cfg
  on target lab system requiring multi-drivers-switch=cvl-2.54:
    - Using subcloud install-value
      extra_boot_params: multi-drivers-switch=cvl-2.54,
      verify that the subcloud switches to the legacy kernel modules
      and the subcloud is able to properly configure its IP address and
      perform the ostree pull operation from the system controller.
- Install subcloud with no extra_boot_params, verify that the
  bootstrap_address_prefix is properly applied. Verify no regression.

Closes-Bug: 2023407

Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I4f3d8e2f240f2aa061de30014cf39dfb9b42a035