Prestaged ISO: copy ostree_repo to versioned platform-backup
This commit applies to the prestaged ISO install. The kickstart.cfg is
updated to copy the prestaged ostree_repo into release-specific
/opt/platform-backup/<release> location.

A minor change is also included in miniboot.cfg to sync the patching
metadata for prepatched ISOs. This fills a potential hole in the
patching metadata sync behaviour identified during testing.
Normally the patching metadata is synchronized from the system
controller down to the subcloud. For the prestaged ISO case, this change
is necessary to ensure the patching metadata is seeded from the
prepatched ISO created via gen-prestaged-iso.sh.

Test Plan
PASS:
- Build prestaged ISO, including container images and a patch
    - Install subcloud using prestaged ISO
    - Verify contents of /opt/platform-backup/<release> are properly
      populated.
    - Verify subcloud is installed using prestaged data from
      /opt/platform-backup/<release>
    - Verify that included container images are installed
- Build prestaged ISO using a pre-patched ISO. Install subcloud, ensure
  that patching metadata is properly synchronized on installation.

Out of scope failure:
- A new bug to be raised for the following:
    - Verify that the included patch is installed on the subcloud
      - It appears that this has never worked in Debian. The --patch
        option makes sense for a Debian installation, since the patches
        are contained in ostree commits. To fully support this
        functionality we need to implement a new mechanism to do a
        sw-patch upload and apply at some point during the installation.
      - Support for the gen-prestaged-iso.sh --patch option will be
        added in a future commit

Closes-Bug: 2039282
Signed-off-by: Kyle MacLeod <kyle.macleod@windriver.com>
Change-Id: I973f4704eae09634a0c3fe2f7fbc31ac1835fcf8