Add LAT hook segment syntax error rendering code comment to kickstarts
The Linux Assembly Tool parses a kickstart file containing various
hook segments that are syntactically delineated with statements that
start with '%'.

The example below shows the start and end syntax of a hook block.

%post --interpreter=/bin/bash

<kickstart bash code goes here>

%end

If LAT detects a hook block syntax error it fails the install and
renders it as a Permission denied like this:

./lat-installer.sh: line 539:
/tmp/lat/lat_post_nochroot_script/10_script: Permission denied

This update adds a comment to the kickstart file header explaining
how LAT renders hook syntax errors as file permission denied and
a comment at the end of the kickstart warning not to remove the
Line Feed at the end of the kickstart files.

Test Plan:

PASS: Verify spelling and format of added code comments.
PASS: Verify kickstart.cfg has a Line Feed at the end of the last
      kickstart.cfg and miniboot.cfg files. They were already there.

Closes-Bug: 2069432
Change-Id: Ic4cd168e784abdba98e45e1837749915db9c4e43
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>