Use new kernel parameter in kickstarts
This commit replaces "multi-drivers-switch" boot parameter with
new "out-of-tree-drivers" boot parameter. As part of "move to in-tree driver"
feature requirement, "multi-drivers-switch" boot parameter is removed as
it is no longer supported and new boot parameter "out-of-tree-drivers"
is added. The new boot parameter was introduced in the
following review:
https://review.opendev.org/c/starlingx/integ/+/922904

Test Plan:
PASS: Verify Build and Install SX system
PASS: Unit test of code block function over an install
PASS: Verify if the out-of-tree-drivers parameter exists on the
      the node install command line then the same option is
      propagated to the disk boot command line.
PASS: USM upgrade on SX from 22.12 to 24.09 -
      Set OOT driver using service parameter before upgrade
      and verify OOT drivers are loaded after upgrade.

Story: 2011056
Task: 50954

Change-Id: Ib4b803423b4204fe661fc2be9bdf1d2e43d1ee47
Signed-off-by: AbhishekJ <abhishek.jaiswal@windriver.com>