Fix change feed ostree config
Sometimes ostree complains about the remaining space on feed
directory while committing, but it has enough space to commit. This
happens because the ostree config to leave some free space is active
and set to a 3% value.

This commit set the min-free-space-percent config to 0, because it
is miscalculating the remaining space and raising an error even with
plenty of free space to commit.

The last commit fixed just part of it:
I149b0c16548369740f5712859725d37ade8ed57a

Test-plan:
PASS: Install a DX system from zero and see if the config was set
PASS: Apply multiple patches and see no error regarding free memory

Story: 2010676
Task: 51128

Change-Id: I68197eed84b1a3cfc034123b897934d40d054613
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>