Add --no-gpg-verify flag to all ostree repos
The way we build pre-patched ISOs has changed, and they are not
GPG-signed anymore. (The GPG validation feature will be enabled in
the future)
Because of this, a fresh installation of a pre-patched ISO was
failing to create the feed repo.

This commit adds the --no-gpg-verify flag in all ostree repos created
during a fresh install. Also, logs if some error happens during an
ostree pull.

Test-Plan:
PASS: Install a 24.09.1 ISO with GPG disabled
      and verify the feed repo is consistent
PASS: Apply a 24.09.2 patch with success
PASS: Install a 24.09.1 ISO with GPG enabled must fail

Story: 2010676
Task: 51534

Change-Id: If1beefbcc78272de76f192a0804170572ab2e7af
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>