Fix miniboot wipe operation
Currently the miniboot wipe operation is reverting the
logic to wipe or not wipe the disk. This is affecting to
the restore operation with install option.

This commit is fixing the logic to wipe or not wipe,
avoiding the wipe disk if wipe_osds parameter is not
defined or zero.

Test Plan:
PASS - Bring up a System Controller with master load
     - Bring up a subcloud with master load without
       wipe_osds parameter in install-values (default)
     - Perform a central backup of the subcloud
     - Perform a restore of the subcloud
     - Verify that the restore operation is successful

PASS - Bring up a subcloud by setting wipe_osds to true
       in install-values.
     - Perform a central backup of the subcloud
     - Perform a restore of the subcloud
     - Verify that the restore operation is successful

PASS - Bring up a subcloud by setting wipe_osds to false
       in install-values.
     - Perform a central backup of the subcloud
     - Perform a restore of the subcloud
     - Verify that the restore operation is successful

Closes-bug: 2111524

Change-Id: If82e49dee119b60d1d0156582b4ea4c4092bd198
Signed-off-by: Cristian Mondo <cristian.mondo@windriver.com>