Add rollback entry to grub.cfg
This commit adds the rollback menu entry on grub, so that the
system can be booted using the previous ostree deployment.

A new boot_failure flag is created, to allow automatic fallback
to rollback deployment in case of a boot loader failure, and a
small fix is implemented to correctly get the kernel_rollback
value from the previous deployment.

Note: initially, this rollback method is designed to be used
      in conjuction with [1]
Note: the automatic management of the boot_failure flag is still
      in progress, so the Test Plan of this commit will cover
      only manual scenarios

[1] https://review.opendev.org/c/starlingx/update/+/950988

Test Plan
PASS: verify the new menu entry appears on grub, with the
      kernel_rollback variable containing the correct kernel
      version instead of the default generic value
PASS: leave boot_failure=0 and verify the system boots on the
      current deployment
PASS: set boot_failure=1 on boot.env and verify the system boots
      in the rollback deployment by default
PASS: AIO-SX stx-10 to stx-11 upgrade:
      - deploy host with to-release
      - after the reboot, set boot_failure=1 and reboot, to
        simulate a boot failure
      - verify system boots from the rollback deployment
      - verify the service from [1] executes and restores the
        system to a from-release state using snapshots
      - when the system boots, verify:
          - system is stable
          - services are enabled-active
          - system and software CLI are usable
          - rollback process can be safely completed

Story: 2011357
Task: 52319

Change-Id: Ie260a15a934625432ac950568a364c2ae2fa43d0
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>