Fail subcloud auto-restore early if missing prestaged data
This commit makes it so the subcloud install during the auto-restore
operation fails early if the local prestaged ostree repo is not found.
During auto-restore, the subcloud will not be able to pull any data
from the system controller, so the subcloud must be prestaged before.

Test Plan:
01. PASS - Run auto-restore without prestaging, check that install
    fails due to timeout and that the log message is in the logs.
02. PASS - Run auto-restore with prestaging, check that install
    succeeds and that it used the prestaged ostree repo, without
    attempting the second ostree pull.
03. PASS - Run standard subcloud restore with install, verify that it
    still works as expected.

Story: 2011454
Task: 52511

Change-Id: Ib849afdb45228b7b1ab5c44238451cb8e8c12544
Signed-off-by: Gustavo Herzmann <gustavo.herzmann@windriver.com>