[Miniboot] Add --no-gpg-verify flag to ostree repos
This commit adds the --no-gpg-verify flag in all subcloud ostree
repos created during a fresh install using miniboot.

Also it set the min-free-space-percent config to 0, because it
is miscalculating the remaining space and raising an error even
with plenty of free space to commit.

Test-Plan:
PASS: Install a subcloud using miniboot with GPG disabled
      and verify the feed repo is consistent
PASS: After, apply patch with success
PASS: Install a subcloud using miniboot
      with GPG enabled must fail

Relates-To: If1beefbcc78272de76f192a0804170572ab2e7af
Relates-To: I149b0c16548369740f5712859725d37ade8ed57a

Closes-Bug: 2119464

Change-Id: I6a45625715138d76b6b2cd04fcae100a68b8e212
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>