mtcAgent: Don't advise enabling standby for AIO Simplex systems
A recent update [1] introduced the posting of a host 'Task' string
that advises to enable a standby controller for a critical active
controller (self) failure on an AIO Simplex system that has
no plan or provision for such standby controller.

[1] https://opendev.org/starlingx/metal/commit/
            2cb728678f7999cfc02bbc088f8fe1545a792754

This update ensures such Task posting does not occur on an
AIO Simplex system.

Test Plan:

PASS: Verify failure of a AIO Simplex system does not post a host
      Task suggestion to provision/enable a standby controller.

Closes-Bug: 2121407
Change-Id: I934ea3385247f1e983a5248129c07903285156b8
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>