Tox/Zuul: Move to starlingx-common jobs
Replacing openstack zuul jobs for starlingx-common jobs.
The main difference, currently, is that the new ones use
the debian-bullseye nodeset, which runs python3.9, while
the openstack one uses ubuntu-noble and python3.12.

This matters because some dependencies and code in the repo
have not yet been updated to run on python3.12, causing errors.

Also, STX is currently based on Bullseye, so this nodeset
better emulates the user's runtime environment.

Test Plan:
pass: Run tox in a debian-bullseye container
pass: Zuul Verification

Closes-Bug: 2130604

Change-Id: I3f7f9128669c3b9cfd140e099f395372f320e1b8
Signed-off-by: Leonardo Fagundes Luz Serrano <Leonardo.FagundesLuzSerrano@windriver.com>